Moving Exceptions into separate files and Minor cleanups.
